requirements.txt
asgiref==3.2.10
astroid==2.4.1
autopep8==1.5.2
boto3==1.17.69
botocore==1.20.69
colorama==0.4.3
Django==3.1
django-cors-headers==3.4.0
django-filter==2.3.0
django-storages==1.11.1
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
rest-framework-generic-relations==2.0.0
s3transfer==0.4.2
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
xlwt==1.3.0
